prog(MKrum): it adds floating point init gradients removing integers

+++ server/src/main/scala/se/kth/rise/bootstrapping/Messages.scala
@@ -27,7 +27,7 @@ import se.sics.kompics.KompicsEvent;
 import se.kth.rise.overlay.Node
 
 case class Boot(assignment: Set[Node]) extends KompicsEvent;
-case class Msg(gradient: List[Int], index : Int) extends KompicsEvent;
+case class Msg(gradient: List[Double], index : Int) extends KompicsEvent;
 case class SharePhase(gradient: Double, index : Int) extends KompicsEvent;
 case object CheckIn extends KompicsEvent;
 case object Ready extends KompicsEvent;
